requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
azure-ai-formrecognizer==3.2.1
azure-common==1.1.28
azure-core==1.26.4
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.0.7
cycler==0.11.0
Django==4.2.1
fonttools==4.39.3
frozenlist==1.3.3
idna==3.4
isodate==0.6.1
joblib==1.2.0
kiwisolver==1.4.4
matplotlib==3.7.1
msrest==0.7.1
multidict==6.0.4
nltk==3.8.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.6
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
plotly==5.14.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
regex==2023.5.5
requests==2.29.0
requests-oauthlib==1.3.1
scikit-learn==1.2.2
scipy==1.10.1
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.2
threadpoolctl==3.1.0
tqdm==4.65.0
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
yarl==1.9.2
django-cors-headers==3.14.0